[#2] feat: iOS part implementation (#4)
* [#2] ios implementation

* [#2] implemented almost all listeners

* [#2] implemented all methods except add layer and add source

* [#2] implemented add layer and add source method with all the supported files

* [#2] refactor codebase

* [#2] upgraded maplibre ios sdk version to `6.11.0`

* [#2] upgraded maplibre android sdk version to 11.8.1

* [#2] applied location settings related args on ios

+++ lib/src/models/light.dart
@@ -10,8 +10,8 @@ class Light {
 
   /// The intensity of the light.
   ///
-  /// This defines how bright the light source is, represented as an integer.
-  final int intensity;
+  /// This defines how bright the light source is, represented as an double.
+  final double intensity;
 
   /// Creates a [Light] instance with the given [color] and [intensity].
   ///
@@ -20,8 +20,8 @@ class Light {
   /// Example:
   /// ```dart
   /// final Light light = Light(
-  ///   color: Color(0xFFFF0000), // Red color
-  ///   intensity: 75,
+  ///   color: #FF0000, // Red color
+  ///   intensity: 0.70,
   /// );
   /// ```
   Light({required this.color, required this.intensity});
